# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Godot + Laravel",
        "description": "API Rest de exemplo para ser utilizada com Godot Engine (https://godotengine.org)",
        "contact": {
            "email": "thiago.bruno@birdy.studio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Insere um usuário",
                "description": "Insere um usuário na API",
                "operationId": "App\\Http\\Controllers\\API\\RegisterController::register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "codename": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "c_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário criado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Autentica um usuário",
                "description": "Autentica um usuário, e retorna os dados de cadastro",
                "operationId": "App\\Http\\Controllers\\API\\RegisterController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "codename": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário autenticado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/forgot_password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Esqueci senha",
                "description": "Envia um código para o usuário conseguir criar nova senha",
                "operationId": "App\\Http\\Controllers\\API\\RegisterController::forgotPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código enviado por email",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/reset_password": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Criar nova senha",
                "description": "Cria uma nova senha com código informado anteriormente",
                "operationId": "App\\Http\\Controllers\\API\\RegisterController::resetPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "c_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código enviado por email",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Retorna o usuário autenticado",
                "description": "Retorna um usuário",
                "operationId": "App\\Http\\Controllers\\API\\UserController::index",
                "responses": {
                    "200": {
                        "description": "Listando o usuário",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Save do usuário",
                "description": "Grava os dados de save do usuário",
                "operationId": "App\\Http\\Controllers\\API\\UserController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "save_data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário salvo com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Remove a conta do usuário",
                "description": "Exclui o usuário e o save de forma permanente",
                "operationId": "App\\Http\\Controllers\\API\\UserController::destroy",
                "responses": {
                    "200": {
                        "description": "Usuário excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/user/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Desconecta a conta do usuário",
                "description": "Efetua o logout do usuário",
                "operationId": "App\\Http\\Controllers\\API\\UserController::logout",
                "responses": {
                    "200": {
                        "description": "Usuário desconectado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "Rotas com funcionalidades do usuário"
        }
    ],
    "externalDocs": {
        "description": "Para ver a explicação sobre o código acesse: https://youtube.com.br/thiagobruno",
        "url": "https://youtube.com.br/thiagobruno"
    }
}